reorganize config pkg wip
