+++ cli/internal/config/file.go
@@ -1,0 +1,130 @@
+package config
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/goccy/go-yaml"
+	"github.com/pkg/errors"
+)
+
+func FromFile(path string, validationType ValidationType) (Config, error) {
+	config := DefaultConfig()
+
+	bz, err := os.ReadFile(path)
+	if err != nil {
+		return Config{}, err
+	}
+
+	// substitute ENV variables
+	expanded := os.ExpandEnv(string(bz))
+
+	opts := []yaml.DecodeOption{
+		yaml.DisallowUnknownField(),
+	}
+
+	if err := yaml.UnmarshalWithOptions([]byte(expanded), &config, opts...); err != nil {
+		return Config{}, err
+	}
+
+	if validationType != ValidationNone {
+		if err := config.Validate2(validationType); err != nil {
+			return Config{}, errors.Wrap(err, "validation failed")
+		}
+	}
+
+	return config, nil
+}
+
+// KeyFileFallbacks returns the paths tried for a local signer key file.
+func KeyFileFallbacks(keyPath string) []string {
+	fallbacks := []string{keyPath}
+
+	// absolute path, no fallbacks needed
+	if filepath.IsAbs(keyPath) {
+		return fallbacks
+	}
+
+	// forgot to add .json extension
+	if !strings.HasSuffix(keyPath, ".json") {
+		keyPath = fmt.Sprintf("%s.json", keyPath)
+
+		fallbacks = append(fallbacks, keyPath)
+	}
+
+	// forgot to add keys/ directory
+	if !strings.Contains(keyPath, "keys/") {
+		keyPath = filepath.Join("keys", keyPath)
+
+		fallbacks = append(fallbacks, keyPath)
+	}
+
+	return fallbacks
+}
+
+func fileExistsInAny(path ...string) error {
+	for _, p := range path {
+		if err := fileExists(p); err == nil {
+			return nil
+		}
+	}
+
+	return errors.New("file not found")
+}
+
+func fileExists(path string) error {
+	info, err := os.Stat(path)
+	if err != nil {
+		return err
+	}
+
+	if info.IsDir() {
+		return errors.Errorf("path is a directory")
+	}
+
+	return nil
+}
+
+func storeConfig(c Config, path string, comments map[string]string) error {
+	if err := EnsureDirectory(path); err != nil {
+		return err
+	}
+
+	bz, err := yaml.MarshalWithOptions(c, yaml.WithComment(toCommentMap(comments)))
+	if err != nil {
+		return err
+	}
+
+	mode := os.FileMode(0o644)
+	if info, statErr := os.Stat(path); statErr == nil {
+		mode = info.Mode().Perm()
+		path, err = filepath.EvalSymlinks(path)
+		if err != nil {
+			return err
+		}
+	} else if !os.IsNotExist(statErr) {
+		return statErr
+	}
+
+	tmp, err := os.CreateTemp(filepath.Dir(path), ".config-*")
+	if err != nil {
+		return err
+	}
+	defer func() { _ = os.Remove(tmp.Name()) }()
+
+	if err := tmp.Chmod(mode); err != nil {
+		_ = tmp.Close()
+		return err
+	}
+	if _, err := tmp.Write(bz); err != nil {
+		_ = tmp.Close()
+		return err
+	}
+	if err := tmp.Close(); err != nil {
+		return err
+	}
+
+	return os.Rename(tmp.Name(), path)
+}
